refactor to async collector class
