+++ django_project/gap/ingestor/tomorrowio/__init__.py
@@ -5,7 +5,7 @@ Tomorrow Now GAP.
 .. note:: Tomorrow.io Ingestors
 """
 
-from django_project.gap.ingestor.tomorrowio.collector import *
+from gap.ingestor.tomorrowio.collector import *
 from gap.ingestor.tomorrowio.duckdb_ingestor import *
 from gap.ingestor.tomorrowio.json_collector import *
 from gap.ingestor.tomorrowio.json_ingestor import *
